docker/requirements.txt
google-cloud-storage==2.19.0
google-cloud-bigquery==3.30.0
pandas==2.2.3
pyarrow==18.1.0
requests==2.32.3
